feat: refactor association endpoints methods, make auth client colocated in task master service
